backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
pinecone
python-dotenv
